Extract the name prompt logic in a helper file
The goal is to allow passing a custom name with custom validation rules.
We'll eventually extract this into it's own package.
